move getSqlConditionFromJSON to helpersUtils
