+++ src/components/permissions/helpersUtils.ts
@@ -1,5 +1,9 @@
 import { PermissionRow, RuleTypes } from './types'
 
+// Need to use require as there are no types for json-sql-builder2, can attempt to add them in the future
+const SQLBuilder = require('json-sql-builder2')
+const jsonToSql = new SQLBuilder('PostgreSQL')
+
 // Constants for compileRowLevelPolicy
 // https://www.postgresql.org/docs/current/sql-createpolicy.html
 const compileRowLevelPolicyRuleTypes: RuleTypes = {
@@ -48,9 +52,31 @@ const remapObjectKeysWithPrefix = (prefix: string, object: Object) => {
   return remappedObject
 }
 
+// Creates SQL from JSON, as per https://github.com/planetarydev/json-sql-builder2, but only for WHERE clause content
+// in { job_title: { $in: ['Sales Manager', 'Account Manager'] }, country_code: 'RU' }
+// out `job_title IN ('Sales Manager', 'Account Manager') AND country_code = 'RU'`
+const getSqlConditionFromJSON = (jsonSQLcondition: any) => {
+  // with no select field specified will default to SELECT *
+  const { sql, values } = jsonToSql.$select({
+    $where: {
+      ...jsonSQLcondition,
+    },
+  })
+
+  let query = sql
+  // $select return key values, which need to be replaced
+  // count backwards in case ${x} above 9 exists
+  for (let i = values.length - 1; i >= 0; i--) {
+    query = query.replace(`$${i + 1}`, values[i])
+  }
+  // Finally return only the WHERE part
+  return query.replace('SELECT * WHERE ', '')
+}
+
 export {
   getPermissionNameAbbreviation,
   getTemplatePermissionAbbreviation,
   remapObjectKeysWithPrefix,
   compileRowLevelPolicyRuleTypes,
+  getSqlConditionFromJSON,
 }
