+++ src/components/permissions/rowLevelPolicyHelpers.ts
@@ -5,10 +5,8 @@ import {
   getTemplatePermissionAbbreviation,
   remapObjectKeysWithPrefix,
   compileRowLevelPolicyRuleTypes,
+  getSqlConditionFromJSON,
 } from './helpersUtils'
-
-const SQLBuilder = require('json-sql-builder2')
-const jsonToSql = new SQLBuilder('PostgreSQL')
 
 /* Compiles JWT from userInfo and PerissionRows
   in { userId: 1, ... }, [
@@ -233,27 +231,6 @@ const compileRowLevelPolicy = (
 
   // Combine them all
   return `${create} ${onFor} ${getUsingClause()} ${withCheck}`
-}
-
-// Creates SQL from JSON, as per https://github.com/planetarydev/json-sql-builder2, but only for WHERE clause content
-// in { job_title: { $in: ['Sales Manager', 'Account Manager'] }, country_code: 'RU' }
-// out `job_title IN ('Sales Manager', 'Account Manager') AND country_code = 'RU'`
-const getSqlConditionFromJSON = (jsonSQLcondition: any) => {
-  // with no select field specified will default to SELECT *
-  const { sql, values } = jsonToSql.$select({
-    $where: {
-      ...jsonSQLcondition,
-    },
-  })
-
-  let query = sql
-  // $select return key values, which need to be replaced
-  // count backwards in case ${x} above 9 exists
-  for (let i = values.length - 1; i >= 0; i--) {
-    query = query.replace(`$${i + 1}`, values[i])
-  }
-  // Finally return only the WHERE part
-  return query.replace('SELECT * WHERE ', '')
 }
 
 // Replaces 'placeholder' templated values in a string
